python-consumers/requirements-test.txt
pytest==7.4.0
requests-mock==1.11.0
psutil==5.9.5
kafka-python==2.0.2
requests==2.31.0
unittest-mock==1.0.1